Added the TextBox class and its drawing and keyboard event handling functions; updated the memory management function parameter types; implemented the log2 function; and added a delay to SVGA redraw.

+++ include/memorymanager.h
@@ -23,7 +23,7 @@ namespace myos {
     private:
         MemoryChunk* first;
     };
-    void *memcpy(void * _dest, void *_src, uint32_t _n);
+    void *memcpy(void * _dest, void *_src, size_t _n);
 }
 
 void* operator new(size_t size);
